ENH: Add some TDT triggering
